+++ Sources/SwiftUIKit/Extensions/View+Conditionals.swift
@@ -19,7 +19,7 @@ public extension View {
             self
         }
     }
-
+    
     /// Make the view searchable if the condition is `true`.
     @ViewBuilder
     func searchable(
@@ -37,6 +37,30 @@ public extension View {
             self
         }
     }
+    
+    #if !os(watchOS) && !os(tvOS)
+    /// Make the view searchable if the condition is `true`.
+    @available(iOS 17.0, macOS 14.0, tvOS 17.0, *)
+    @ViewBuilder
+    func searchable(
+        if condition: Bool,
+        text: Binding<String>,
+        isPresented: Binding<Bool>,
+        placement: SearchFieldPlacement = .automatic,
+        prompt: Text? = nil
+    ) -> some View {
+        if condition {
+            self.searchable(
+                text: text,
+                isPresented: isPresented,
+                placement: placement,
+                prompt: prompt
+            )
+        } else {
+            self
+        }
+    }
+    #endif
     
     /// Show the view if the provided condition is `true`.
     func visible(if condition: Bool) -> some View {